+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [11.4.0] - 2023-01-13
 - Disable Gravity Forms update messages and background updates
 - Flush Kinsta cache rather than WP-Rocket when present
-- Only force Mailhog in local environments
+- Don't force mailers
 
 ## [11.3.1] - 2022-12-15
 - Define Kinsta mu-plugin URL